backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "aiofiles>=24.1.0",
    "boto3>=1.38.35",
    "celery[redis]>=5.3.1",
    "faiss-cpu>=1.7.4",
    "fastapi[all,standard]>=0.115.12",
    "filetype>=1.2.0",
    "groq>=0.26.0",
    "langchain>=0.3.25",
    "langchain-community>=0.3.25",
    "langchain-google-genai>=2.1.5",
    "langgraph>=0.4.8",
    "pillow>=11.2.1",
    "pypdf>=5.6.0",
    "python-docx>=1.2.0",
    "python-dotenv>=1.1.0",
    "python-magic>=0.4.27",
    "python-multipart>=0.0.20",
    "redis>=6.2.0",
    "requests>=2.32.4",
    "sqlalchemy>=2.0.0",
    "tiktoken>=0.9.0",
    "websockets>=15.0.1",
]
